\name{Drugs.indata}
\alias{Phenotypes}
\docType{data}
\title{
Drug data
}
\description{
Drug response information 
}
\usage{data("Cv")}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Etoposide}}{a numeric vector}
    \item{\code{Fascaplysin}}{a numeric vector}
    \item{\code{Bortezomib}}{a numeric vector}
    \item{\code{Geldanamycin}}{a numeric vector}
  }
}
\details{
Drug response information 
}
\source{
A next generation tool enables identification of functional cancer subtypes with associated biological phenotypes.
}
\references{
Nyamundanda et al (2015). A next generation tool enables identification of functional cancer subtypes with associated biological phenotypes.
}
\examples{
data(Cv)
## maybe str(Cv) ; plot(Cv) ...
}
\keyword{datasets}
