% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emerg.R
\name{emerg}
\alias{emerg}
\title{Mapping emergence date of insect after overwintering}
\usage{
emerg(DVD0 = NULL, GDD = NULL, tasmax, tasmin)
}
\arguments{
\item{DVD0}{basal temperature/low threshold development temperature, based on which insect start to develop}

\item{GDD}{growing degree days of egg to adult}

\item{tasmax}{stack raster of grid daily maximum temperature, SpatRaster object (1-365 days basis)}

\item{tasmin}{stack raster of grid daily minimum temperature, SpatRaster object (1-365 days basis)}
}
\description{
Mapping emergence date of insect after overwintering
}
