% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen.R
\name{Gen}
\alias{Gen}
\title{Mapping number of generation based on basal temperature and growing degree days}
\usage{
Gen(DVD0 = NULL, GDD = NULL, stk)
}
\arguments{
\item{DVD0}{basal temperature/low threshold development temperature, based on which insect start to develop}

\item{GDD}{growing degree days of egg to adult}

\item{stk}{stack raster of grid daily mean temperature, SpatRaster object (1-365 days basis)}
}
\description{
Mapping number of generation based on basal temperature and growing degree days
}
