\name{PhenMAP-package}
\alias{PhenMAP-package}
\alias{PhenMAP}
\docType{package}
\title{
\packageTitle{PhenMAP}
}
\description{
\packageDescription{PhenMAP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PhenMAP}
\packageIndices{PhenMAP}
A statistical tool for clustering expression data, finding associated phenotypes and molecular features.
}
\author{
\packageAuthor{PhenMAP}

Maintainer: \packageMaintainer{PhenMAP}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
